refactor(playground): split robota executor module
