+++ packages/agent-playground/src/lib/playground/robota-executor/constants.ts
@@ -1,0 +1,5 @@
+export const REMOTE_EXECUTOR_TIMEOUT_MS = 30000;
+export const HISTORY_MAX_EVENTS = 1000;
+export const STATISTICS_MAX_ENTRIES = 1000;
+export const SLOW_EXECUTION_THRESHOLD_MS = 3000;
+export const ERROR_RATE_THRESHOLD = 10;
