+++ packages/agent-playground/src/lib/playground/robota-executor/index.ts
@@ -1,0 +1,13 @@
+export { PlaygroundExecutor } from './playground-executor';
+
+export type {
+  IConversationEvent,
+  IPlaygroundAgentConfig,
+  IPlaygroundExecutorResult,
+  IPlaygroundPlugin,
+  IPlaygroundTool,
+  IPlaygroundUiError,
+  IVisualizationData,
+  TPlaygroundMode,
+  TPlaygroundUiErrorKind,
+} from '../robota-executor-types';
