+++ packages/agent-playground/src/lib/playground/robota-executor/types.ts
@@ -1,0 +1,12 @@
+export interface IAgentConfigurationSnapshot {
+  version: number;
+  tools: Array<{ name: string; parameters?: string[] }>;
+  updatedAt: number;
+  metadata?: Record<string, string | number | boolean | Date | string[]>;
+}
+
+export interface IToolCard {
+  id: string;
+  name: string;
+  description?: string;
+}
